Introduce standalone `invokeGraph` function and switch `runOnce` to call it.
