+++ packages/breadboard/src/runner.ts
@@ -30,7 +30,7 @@ import {
 import { createLoader } from "./loader/index.js";
 import { GraphLoader, GraphProvider } from "./loader/types.js";
 import { toMermaid } from "./mermaid.js";
-import { timestamp } from "./timestamp.js";
+import { invokeGraph } from "./run/invoke-graph.js";
 import { runGraph } from "./run/run-graph.js";
 
 /**
@@ -120,47 +120,7 @@ export class BoardRunner implements BreadboardRunner {
     inputs: InputValues,
     context: NodeHandlerContext = {}
   ): Promise<OutputValues> {
-    const args = { ...inputs, ...this.args };
-    const { probe } = context;
-
-    try {
-      let outputs: OutputValues = {};
-
-      const path = context.invocationPath || [];
-
-      for await (const result of this.run(context)) {
-        if (result.type === "input") {
-          // Pass the inputs to the board. If there are inputs bound to the
-          // board (e.g. from a lambda node that had incoming wires), they will
-          // overwrite supplied inputs.
-          result.inputs = args;
-        } else if (result.type === "output") {
-          outputs = result.outputs;
-          // Exit once we receive the first output.
-          await probe?.report?.({
-            type: "nodeend",
-            data: {
-              node: result.node,
-              inputs: result.inputs,
-              outputs,
-              path: [...path, result.invocationId],
-              timestamp: timestamp(),
-            },
-          });
-          await probe?.report?.({
-            type: "graphend",
-            data: { path, timestamp: timestamp() },
-          });
-          break;
-        }
-      }
-      return outputs;
-    } catch (e) {
-      // Unwrap unhandled error (handled errors are just outputs of the board!)
-      if ((e as Error).cause)
-        return { $error: (e as Error).cause } as OutputValues;
-      else throw e;
-    }
+    return invokeGraph(this, inputs, context);
   }
 
   /**
